app/page refac
